industry2018 move to common structure
