+++ src/climatevision/generator/industry2018/energy_source/__init__.py
@@ -2,8 +2,8 @@
 
 from dataclasses import dataclass
 
-from ..inputs import Inputs
-from ..common.energy import Energy, EnergyWithPercentage
+from ...inputs import Inputs
+from ...common.energy import Energy, EnergyWithPercentage
 
 
 @dataclass(kw_only=True)
